Combine chart templates for regular and baking nodes.
We do this by using {{ range }} to iterate over .Values.nodes.
Currently, this contains baking and regular.  We change the rest
of the template infrastructure to adapt to the new methodology.

We used ./bin/helm-compare-versions to validate that (at least
for our values.yaml file) we did not make any changes to the
generated charts.

This commit is in preparation for a slightly more wide ranging
change which will allow users to define their own types of nodes
which will include optional containers.